requirements.txt
alembic==1.9.2
attrs==23.1.0
bidict==0.22.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Faker==18.4.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.2
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
funcs==0.8.0
future==0.18.3
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mysqlclient==2.2.1
named==1.3.0
packaging==23.1
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.4.1
python-socketio==5.8.0
ratelim==0.1.6
requests==2.31.0
simple-websocket==0.10.0
six==1.16.0
solus==1.1.0
SQLAlchemy==1.4.46
typing-aliases==1.3.1
typing_extensions==4.9.0
urllib3==2.0.2
Werkzeug==2.2.2
wraps==0.6.1
wsproto==1.2.0
WTForms==3.0.1
zipp==3.15.0
